Update 9 Nov

+++ cms-new/admin/includes/view_all_posts.php
@@ -23,6 +23,29 @@ if (isset($_POST['checkBoxArray'])) {
         confirmQuery($update_to_delete_status);
         break;
 
+        case 'clone':
+        $query = "SELECT * FROM posts WHERE post_id = {$postValueId} ";
+        $select_post_query = mysqli_query($connection, $query);
+
+        while ($row = mysqli_fetch_array($select_post_query)) {
+            $post_title = $row['post_title'];
+            $post_category_id = $row['post_category_id'];
+            $post_date = $row['post_date'];
+            $post_author = $row['post_author'];
+            $post_status = $row['post_status'];
+            $post_image = $row['post_image'];
+            $post_tags = $row['post_tags'];
+            $post_content = $row['post_content'];
+        }
+
+        $query = "INSERT INTO posts(post_title, post_category_id, post_date, post_author, post_status, post_image, post_tags, post_content) ";
+        $query .= "VALUES ('$post_title', $post_category_id, now(), '$post_author', '$post_status', '$post_image', '$post_tags', '$post_content') ";
+        $copy_query = mysqli_query($connection, $query);
+
+        if (!$copy_query) {
+            die("QUERY FAILED" . mysqli_error($connection));
+        }
+        break;
     }    
     }   
 }
@@ -40,6 +63,7 @@ if (isset($_POST['checkBoxArray'])) {
         <option value="published">Publish</option>    
         <option value="draft">Draft</option>    
         <option value="delete">Delete</option>    
+        <option value="clone">Clone</option>    
     </select>
 </div>
 
@@ -62,12 +86,13 @@ if (isset($_POST['checkBoxArray'])) {
                 <th>Date</th>
                 <th>Edit</th>
                 <th>Delete</th>
+                <th>Views</th>
             </tr>
         </thead>
         <tbody>
 
 <?php // prikaz postova u posts.php
-$query = "SELECT * FROM posts"; 
+$query = "SELECT * FROM posts ORDER BY post_id DESC"; 
 $select_posts = mysqli_query($connection, $query);
 
 while ($row = mysqli_fetch_assoc($select_posts)) {
@@ -80,6 +105,7 @@ while ($row = mysqli_fetch_assoc($select_posts)) {
     $post_tags = $row['post_tags'];
     $post_comment_count = $row['post_comment_count'];
     $post_date = $row['post_date'];
+    $post_views_count = $row['post_views_count'];
 
     echo "<tr>";
     ?>
@@ -109,7 +135,8 @@ while ($row = mysqli_fetch_assoc($select_categories_id)) {
     echo "<td>{$post_comment_count}</td>";
     echo "<td>{$post_date}</td>";
     echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-    echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+    echo "<td><a onClick=\"javascript: return confirm('Are you sure?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
+    echo "<td><a href='posts.php?reset={$post_id}'>$post_views_count</a></td>";
     echo "</tr>";
 }
 ?>
@@ -126,4 +153,11 @@ if (isset($_GET['delete'])) {
     header("Location: ./posts.php"); // mora refreshati stranu da se vidi bez deletanog
 }
 
+if (isset($_GET['reset'])) {  // ako klikne broj pregleda resetaj broj pregleda na 0
+    $the_post_id = $_GET['reset'];
+
+    $query = "UPDATE posts SET post_views_count = 0 WHERE post_id =". mysqli_real_escape_string($connection, $the_post_id) . " ";
+    $reset_query = mysqli_query($connection, $query); 
+    header("Location: ./posts.php"); // mora refreshati stranu da se vidi bez deletanog
+}
 ?>
